Turret now pivots with the right joystick and faces the direction the stick is facing.
The movement of the joystick is almost overly precise. I modified the sensitivity but it may need more work via Lerp, though my first attempt using Lerp to smooth the Input failed.

The player and player projectiles are now on their own layers and no longer collide with each other. This is essentially turning friendly fire off. So that's good.