Adding mission framework.
Some TODOs are missing -- mainly I'm not 100% sure how we should expand on this, I think it might be best to just add in elements as we go and make changes as needed. An example of what still needs to be added would be an element that defines what enemy type to spawn and how many of that specific type, as well as their difficulty.

https://app.asana.com/0/19050287655194/19955992770458